Update README.md
I was going to publish this fork to the KDE Store now, but I just
noticed that another person had already done the same thing just
a few hours (or less) ago... The method for determining the primary
monitor is quite different, though.

Now I need to reconsider what to do.